+++ tedi/components/form/date-field/date-field-modal/date-field-modal.component.spec.ts
@@ -36,6 +36,8 @@ function makeData(
     unavailableDays: undefined,
     shouldDisableMonth: undefined,
     shouldDisableYear: undefined,
+    minYear: null,
+    maxYear: null,
     closeOnSelect: true,
     ...overrides,
   };